ragent/requirements.txt
fastapi==0.115.13
uvicorn==0.23.2
aiohttp==3.8.5
beautifulsoup4==4.12.2
selenium==4.11.2
undetected-chromedriver==3.5.3
nest_asyncio==1.6.0

fastapi-limiter

langchain==0.3.26
langchain-google-genai==2.1.5 
langchain-core==0.3.66
langgraph==0.4.9



python-dotenv==1.0.0
tenacity==8.2.3
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.0
structlog==23.1.0
asyncpraw
asyncprawcore
python-jose[cryptography]==3.3.0  # for JWT support
passlib[bcrypt]==1.7.4      
psycopg2-binary==2.9.9 

# Celery and its dependencies
celery==5.3.6
flower==2.0.1  # Optional: for monitoring Celery tasks 

# Pydantic settings
pydantic-settings==2.10.1

rpdb==0.1.6

# Twitter integration
twikit==2.3.3

# Search APIs
exa-py==1.14.6  # Exa API client
tavily-python==0.3.1  # Tavily API client

# Vector Database and Embeddings
qdrant-client==1.7.0  # Vector database client
# Commented out SentenceTransformer dependencies
# sentence-transformers==2.2.2  # For generating text embeddings
# torch>=2.0.0  # Required by sentence-transformers
# numpy>=1.24.0  # Required for vector operations

# huggingface-hub==0.16.4
# Google
google-genai==1.20

# ---- New dependencies for async rate-limiting ----
aiolimiter==1.1.0

#-----Testing packages--------
pytest
pytest-asyncio

# --------Authenitication and Email-------
email-validator
fastapi-mail==1.5.0
Jinja2==3.1.6
PyJWT==2.10.1
python-multipart==0.0.20
Authlib==1.6.0
itsdangerous==2.2.0
